Правка сборки
