+++ app/src/main/java/ru/practicum/android/diploma/di/InteractorModule.kt
@@ -3,6 +3,7 @@ package ru.practicum.android.diploma.di
 import org.koin.dsl.module
 import ru.practicum.android.diploma.domain.areas.AreasInteractor
 import ru.practicum.android.diploma.domain.areas.impl.AreasInteractorImpl
+import ru.practicum.android.diploma.domain.countrychoice.CountryUseCase
 import ru.practicum.android.diploma.domain.favorite.FavoriteInteractor
 import ru.practicum.android.diploma.domain.favorite.FavoriteInteractorImpl
 import ru.practicum.android.diploma.domain.filteringsettings.FilterInteractor
@@ -50,5 +51,9 @@ val interactorModule = module {
 
     factory<IndustryInteractor> {
         IndustryInteractorImpl(get(), get())
+    }
+
+    factory {
+        CountryUseCase(get())
     }
 }
