+++ app/src/main/java/ru/practicum/android/diploma/domain/areas/AreasInteractor.kt
@@ -5,5 +5,4 @@ import ru.practicum.android.diploma.util.Resource
 
 interface AreasInteractor {
     suspend fun getAreas(): Resource<List<Area>>
-    suspend fun getCountries(): Resource<List<Area>>
 }
