+++ app/src/main/java/ru/practicum/android/diploma/presentation/countrychoice/CountryViewModel.kt
@@ -5,12 +5,12 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.launch
-import ru.practicum.android.diploma.domain.areas.AreasInteractor
+import ru.practicum.android.diploma.domain.countrychoice.CountryUseCase
 import ru.practicum.android.diploma.presentation.countrychoice.models.CountryScreenState
 import ru.practicum.android.diploma.util.Resource
 
 class CountryViewModel(
-    val interactor: AreasInteractor
+    val useCase: CountryUseCase
 ) : ViewModel() {
     private val screenState = MutableLiveData<CountryScreenState>()
     fun getScreenState(): LiveData<CountryScreenState> = screenState
@@ -21,7 +21,7 @@ class CountryViewModel(
 
     private fun loadCountries() {
         viewModelScope.launch {
-            when (val result = interactor.getCountries()) {
+            when (val result = useCase.getCountries()) {
                 is Resource.Success -> {
                     result.data?.let { countries ->
                         val sortedCountries = countries.sortedBy { it.name }
